trekker ut den private kanSnikeIKøen-funksjonen fra AutovedtakSatsendringRollbackService og legger den inn som en public gjenbrukbar funksjon i SnikeIkøenService,
